requirements.txt
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
pyyaml>=6.0
tqdm>=4.66
regex>=2023.12.25

# ingest
requests>=2.31
beautifulsoup4>=4.12

# demo
streamlit>=1.30

# transformer
torch>=2.1
transformers>=4.41
datasets>=2.20
accelerate>=0.30
